# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_smart-invest-solutions_backend_internal_domain.AccessRequest:
    properties:
      admin_notes:
        type: string
      applied_agency_id:
        description: |-
          AppliedAgencyID is the AdminID (e.g. "ADM-7F3K9Q") the applicant supplied to say which
          agency/admin they want to be managed by. Validated against a real admin account at
          submission time. Empty means unassigned — visible only to a super_admin.
        type: string
      applied_referral_code:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      notes:
        type: string
      phone:
        type: string
      status:
        description: PENDING, APPROVED, REJECTED
        type: string
      updated_at:
        type: string
    required:
    - email
    - name
    - phone
    type: object
...
  github_com_smart-invest-solutions_backend_pkg_response.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      message:
        type: string
      page:
        type: integer
      success:
        type: boolean
      total:
        type: integer
      total_pages:
        type: integer
    type: object
info:
  contact:
    email: support@ulmind.com
    name: ULMIND Team
  description: |-
    Backend API for Smart Invest Solutions — a LIC Policy & Investment Management Platform.
    This API provides user registration, authentication with JWT, role-based access control, and more.
  title: Smart Invest Solutions API
  version: "1.0"
paths:
  /access-requests:
    get:
      consumes:
      - application/json
      description: Retrieves access requests with optional filtering by status (PENDING,
        APPROVED, REJECTED). A super_admin sees every request; a plain admin sees
        only requests whose Agency ID matches their own Admin ID.
      parameters:
      - description: 'Filter by status: PENDING, APPROVED, REJECTED'
        in: query
        name: status
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Access requests retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.AccessRequest'
                  type: array
              type: object
        "401":
          description: Unauthorized — Bearer Token required
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "403":
          description: Forbidden — Admin role required
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all Access Requests (Admin Only)
      tags:
      - Access Requests (Admin)
    post:
      consumes:
      - application/json
      description: Submits a request for access to the platform with Name, Contact
        No, and Email. An optional Agency ID (an admin's Admin ID) routes the request
        to that admin specifically — an invalid Agency ID is rejected. Status is set
        to PENDING until approved by an Admin.
      parameters:
      - description: Access Request Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.CreateAccessRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Access request submitted successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.AccessRequest'
              type: object
        "400":
          description: Bad request (email already exists or pending request exists)
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      summary: Request for Access (Client Sign up request)
      tags:
      - Access Requests
  /access-requests/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves detailed info for a single access request.
      parameters:
      - description: Access Request ID (MongoDB ObjectID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Access request retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.AccessRequest'
              type: object
        "404":
          description: Request not found
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Access Request by ID (Admin Only)
      tags:
      - Access Requests (Admin)
  /access-requests/{id}/approve:
    post:
      consumes:
      - application/json
      description: Approves a client's access request. Automatically generates a secure
        random password, creates a User account, updates request status to APPROVED,
        and emails login credentials to the client via Resend.
      parameters:
      - description: Access Request ID
        in: path
        name: id
        required: true
        type: string
      - description: Optional Admin Notes
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.ApproveAccessRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Access request approved, account created, and credentials emailed
          schema:
            allOf:
            - $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.UserResponse'
              type: object
        "400":
          description: Bad request (already approved or invalid ID)
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "403":
          description: Forbidden — Admin role required
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Approve Access Request & Email Credentials (Admin Only)
      tags:
      - Access Requests (Admin)
  /access-requests/{id}/reject:
    post:
      consumes:
      - application/json
      description: Rejects a client's access request and marks status as REJECTED.
      parameters:
      - description: Access Request ID
        in: path
        name: id
        required: true
        type: string
      - description: Optional Rejection Reason
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.RejectAccessRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Access request rejected
          schema:
            allOf:
            - $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.AccessRequest'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "403":
          description: Forbidden — Admin role required
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Reject Access Request (Admin Only)
      tags:
      - Access Requests (Admin)
  /admins:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of all admin and super_admin accounts.
        Only accessible by super_admin.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Admin accounts retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.UserResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "403":
          description: Forbidden — super_admin role required
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all Admin accounts (Super Admin only)
      tags:
      - Admin Accounts
    post:
      consumes:
      - application/json
      description: Creates a new Admin account from Name/Email/Phone/ExpiryDate. Auto-generates
        a unique Admin ID, a random password, and a 4-digit PIN, then emails the credentials
        to the new admin. ExpiryDate must be in the future — once it passes, the admin
        cannot log in until a Super Admin renews it via PUT /admins/{id}/expiry. The
        response also returns the plaintext password/PIN once, so the Super Admin
        can share them even if the email fails to deliver. Only accessible by super_admin.
      parameters:
      - description: New Admin details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.CreateAdminRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Admin account created successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.CreateAdminResponse'
              type: object
        "400":
          description: Bad request (e.g. email already in use)
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "403":
          description: Forbidden — super_admin role required
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create Admin account (Super Admin only)
      tags:
      - Admin Accounts
  /admins/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently deletes an admin account and its associated data. Self-deletion
        is disallowed — use the account settings delete option instead. Only accessible
        by super_admin.
      parameters:
      - description: Admin User ID (MongoDB ObjectID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Admin account deleted successfully
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "403":
          description: Forbidden — super_admin role required
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete Admin account (Super Admin only)
      tags:
      - Admin Accounts
  /admins/{id}/expiry:
    put:
      consumes:
      - application/json
      description: Sets a new (future) expiry date on an admin account, reactivating
        login for an already-expired admin, and emails the admin a confirmation. Only
        accessible by super_admin.
      parameters:
      - description: Admin User ID (MongoDB ObjectID)
        in: path
        name: id
        required: true
        type: string
      - description: New expiry date
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.RenewAdminExpiryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Admin expiry date renewed successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.UserResponse'
              type: object
        "400":
          description: Bad request (e.g. expiry date not in the future, or target
            is not an admin account)
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "403":
          description: Forbidden — super_admin role required
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Renew an admin account's expiry date (Super Admin only)
      tags:
      - Admin Accounts
...
  /users/forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a 6-digit OTP code to the registered email address if an
        account exists.
      parameters:
      - description: Registered Email Address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_smart-invest-solutions_backend_internal_domain.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: If an account exists with this email, an OTP has been sent.
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_smart-invest-solutions_backend_pkg_response.APIResponse'
      summary: Request password reset OTP
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your Bearer token in the format: **Bearer &lt;token&gt;**'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"